requirements.txt
# QF Framework (local wheel — install via: pip install dist/qf-1.0.2-py3-none-any.whl)
# qf @ file:///app/dist/qf-1.0.2-py3-none-any.whl  ← Docker path; use wheel arg locally

# API / framework
flask>=3.0.0
flask-restx>=1.3.0
gevent>=23.9.0
psycogreen>=1.0.2

# QF framework transitive deps
redis>=5.0.0
kafka-python>=2.0.2

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Validation
pydantic>=2.6.0

# Auth / Keycloak
python-jose[cryptography]>=3.3.0
python-keycloak>=4.0.0

# Storage (S3/MinIO)
boto3>=1.34.0

# Observability
prometheus-client>=0.20.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-instrumentation-flask>=0.46b0
opentelemetry-exporter-otlp-proto-grpc>=1.25.0

# Config / utils
python-dotenv>=1.0.0
requests>=2.31.0
colorama>=0.4.0
httpx>=0.27.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-bdd>=7.0.0
responses>=0.25.0
faker>=25.0.0
testcontainers[postgres,redis,kafka]>=4.0.0
